refactor(mm-next): add shared Hint component
